Day 2 solution using recursive int-machine function
The day 2 solution uses a recursive function to implement the int machine. The
function takes the initial instruction ptr and memory as input and recursively
executes instructions until the 99 terminate operation code is encountered. It
then returns the memory where position 0 contains the answer.

The second task is implemented by first constructing tuples for all combinations
of the two ranges 0 - 99. This is done using the for function. The tuples are
then mapped over in parallel using pmap, where each tuple is used as arguments
to the int-machine function. We then drop results until the result equals the
number we are looking for. Finally we return the result of (100 * noun) + verb.

For this task the as-> threading macro is use to make the threading easier
considering the forms should be threaded in different positions for many of the
functions being called.

Signed-off-by: Snorre Magnus Davøen <snorre.davoen@nsd.no>