Day 3 solution based on distance tracking + maps for intersections
The solution generates all coordinates in the two dimensional x y space for each
of the two lines. For each coordinate we keep track of the distance traveled to
make task 2 possible to implement.

The two coordinate vectors can be used to construct two maps which is then easy
to intersect selecting all key value pairs in map 1 matching with keys in map 2.
This gives us the intersection points for task 1 and 2.

For task 1 we use the keys in the intersection points map to calculate the
shortes manhatten distance to coordinate 0,0.

For task 2 we take the distances in the intersection points map and sum them
with distances from map 2 thus getting a map of intersections where the values
are the combined distance traveled to the intersection. We then select the
values and sort.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>