Solution to day 4
Creates predicate functions for each rule in the task. The predicates are
combined and used to filter the numbers in the range provided by the task. Then
the task is solved by counting the filtered sequence of passwords.

The first task finds pairs of numbers using a simple regex with a capturing
group for the numbers 0 to 9, and then a rematch for the most recently captured
group.

The second task finds repeated numbers using partition-by and then counting the
length of sequences. It filters down the groups of repeated digits to the ones
of length 2. The predicate for repeated numbers then simply checks if there is
at least one sequence of repeated numbers of length 2.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>